print in safe mode

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -10,7 +10,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	const listint_t *slow, *fast;
-	int nodes = 0;
+	size_t nodes = 0;
 
 	if (head == NULL)
 		exit(98);
